w00t! straight json is working
By attaching a sparse-checkout of just the textmate bundles plugin from
intellij community repository, I was able to see the exact filenames it
looks for in the Syntaxes folder, and it has a JsonPlistReader in the
plugin.

So - no need to have the awkward steps through Sublime editor or its
python script.